requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
Authlib==1.0.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.1
cryptography==36.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.0
fastapi-sso==0.2.11
frozenlist==1.3.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
rfc3986==1.5.0
rsa==4.8
sendgrid==6.9.7
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starkbank-ecdsa==2.0.3
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.1
websockets==10.2
yarl==1.7.2
